requirements.txt
amqp==5.1.1
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
crispy-bootstrap5==0.6
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==4.0
django-cache-url==3.4.0
django-cleanup==6.0.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
environs==9.5.0
kombu==5.2.4
marshmallow==3.15.0
packaging==21.3
Pillow==9.1.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
pyparsing==3.0.8
python-dotenv==0.20.0
pytz==2022.1
signals==0.0.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
